refactor: split user oidc account routes
